Compare on UInt64 instead of UInt
error: integer literal '4294967295' overflows when stored into 'Int'

+++ Sources/String.swift
@@ -25,7 +25,7 @@ extension String: PackProtocol {
             }
         }
 
-        let n = UInt(bytes.count)
+        let n = UInt64(bytes.count)
 
         if n == 0 {
             return [ 0x80 ]
